@@ -3,7 +3,7 @@ import { BrowserModule }            from '@angular/platform-browser';
 import { RouterModule }             from '@angular/router';
 import { NgModule }                 from '@angular/core';
 import { FormsModule }              from '@angular/forms';
-import { HttpModule }               from '@angular/http';
+import { HttpClientModule }         from '@angular/common/http';
 import { routes }                   from './app.routes';
 import { AppComponent }             from './app.component';
 import { RegisterComponent }        from './register/register.component';
@@ -27,7 +27,7 @@ import { AdminComponent } from './admin/admin.component';
   imports: [
     BrowserModule,
     FormsModule,
-    HttpModule,
+    HttpClientModule,
     RouterModule.forRoot(routes)
   ],
   providers: [UserService],
