@@ -2,7 +2,7 @@ import { Component, OnInit }      from '@angular/core';
 import { UserService }            from '../_services/user.service';
 import { Router }                 from '@angular/router';
 import { LoginComponent }         from '../login/login.component'
-import {BehaviorSubject}          from 'rxjs/BehaviorSubject'
+import {BehaviorSubject}          from 'rxjs'
 
 @Component({
   selector: 'navbar',
